Factored out DeleteProjectDialog class
